Modularize export functionality and make it extensible
