+++ KInspector.Modules/Kentico.KInspector.Modules.csproj
@@ -49,6 +49,14 @@
       <HintPath>..\packages\JSErrorCollector.0.6.1.0\lib\net40\JSErrorCollector.dll</HintPath>
       <Private>True</Private>
     </Reference>
+    <Reference Include="Ninject, Version=3.2.0.0, Culture=neutral, PublicKeyToken=c7192dc5380945e7, processorArchitecture=MSIL">
+      <HintPath>..\packages\Ninject.3.2.2.0\lib\net45-full\Ninject.dll</HintPath>
+      <Private>True</Private>
+    </Reference>
+    <Reference Include="Ninject.Extensions.Conventions, Version=3.2.0.0, Culture=neutral, PublicKeyToken=c7192dc5380945e7, processorArchitecture=MSIL">
+      <HintPath>..\packages\Ninject.Extensions.Conventions.3.2.0.0\lib\net45-full\Ninject.Extensions.Conventions.dll</HintPath>
+      <Private>True</Private>
+    </Reference>
     <Reference Include="NPOI, Version=2.2.1.0, Culture=neutral, PublicKeyToken=0df73ec7942b34e1, processorArchitecture=MSIL">
       <HintPath>..\packages\NPOI.2.2.1\lib\net40\NPOI.dll</HintPath>
       <Private>True</Private>
@@ -85,8 +93,15 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Helpers\ExportExtensions.cs" />
-    <Compile Include="Helpers\ExportHelper.cs" />
+    <Compile Include="Export\ExportModuleLoader.cs" />
+    <Compile Include="Export\ExportModuleMetaData.cs" />
+    <Compile Include="Export\IExportModule.cs" />
+    <Compile Include="Export\Modules\ExportDocx.cs" />
+    <Compile Include="Export\Modules\ExportDocxExtensions.cs" />
+    <Compile Include="Export\Modules\ExportXlsx.cs" />
+    <Compile Include="Export\Modules\ExportXlsxExtensions.cs" />
+    <Compile Include="Export\Modules\ExportXml.cs" />
+    <Compile Include="Export\Modules\ExportXmlExtensions.cs" />
     <Compile Include="Helpers\MacroValidator.cs" />
     <Compile Include="Helpers\ProbeHelper.cs" />
     <Compile Include="Helpers\ProjectCodeFilesHelper.cs" />
@@ -338,7 +353,9 @@
     <Content Include="Data\Templates\KInspectorReportTemplate.docx">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Content>
-    <None Include="packages.config" />
+    <None Include="packages.config">
+      <SubType>Designer</SubType>
+    </None>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Import Project="$(SolutionDir)\.nuget\NuGet.targets" Condition="Exists('$(SolutionDir)\.nuget\NuGet.targets')" />
