+++ KInspector.Modules/packages.config
@@ -1,6 +1,8 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
   <package id="JSErrorCollector" version="0.6.1.0" targetFramework="net45" />
+  <package id="Ninject" version="3.2.2.0" targetFramework="net45" />
+  <package id="Ninject.Extensions.Conventions" version="3.2.0.0" targetFramework="net45" />
   <package id="NPOI" version="2.2.1" targetFramework="net45" />
   <package id="Selenium.Support" version="2.52.0" targetFramework="net45" />
   <package id="Selenium.WebDriver" version="2.52.0" targetFramework="net45" />
