+++ KInspector.Web/FrontEnd/Scripts/Services.js
@@ -166,22 +166,46 @@
         /**
          * Handles export setup and report execution
          */
-        .factory('kiExportService', ['knlTargetConfigService', 'knlModuleService', 'knlErrorService', function (configService, moduleService, errorService) {
+        .factory('kiExportService', ['$http', '$q', 'knlTargetConfigService', 'knlModuleService', 'knlErrorService', function ($http, $q, configService, moduleService, errorService) {
+            // Init cache
+            localStorage.setItem('kiExportServiceExportTypes', null);
+
             return {
                 selectedModules: [],
                 selectorsVisible: false,
-               
+
+                /**
+                 * Return available eport types
+                 */
+                getExportTypes: function() {
+                    var deferred = $q.defer();
+                    var cachedExportTypes = JSON.parse(localStorage.getItem('kiExportServiceExportTypes'));
+
+                    if (cachedExportTypes) {
+                        // Return cached results
+                        deferred.resolve(cachedExportTypes);
+                    } else {
+                        $http.get("http://localhost:9000/api/export/GetExportTypes", { cache: true })
+                            .success(function (exportTypes) {
+                                localStorage.setItem('kiExportServiceExportTypes', JSON.stringify(exportTypes));
+                                deferred.resolve(exportTypes);
+                            })
+                            .error(deferred.error);
+                    }
+                    return deferred.promise;
+                },
+
                 /**
                  * Runs selected modules and returns module results as a file
                  */
-                exportReport: function (exportType) {
+                exportReport: function(exportType) {
                     if (exportType == undefined) {
                         errorService.triggerError("No export type selected");
                         return;
                     }
 
                     var paramsWithModuleNames = angular.extend({ moduleNames: this.selectedModules.sort() }, configService.getConfig(), { exportType: exportType });
-                    var url = "http://localhost:9000/api/modules/GetModulesResults?" + $.param(paramsWithModuleNames);
+                    var url = "http://localhost:9000/api/export/GetModuleExport?" + $.param(paramsWithModuleNames);
 
                     window.open(url, '_blank');
                 }
