+++ KInspector.Web/Kentico.KInspector.Web.csproj
@@ -117,6 +117,7 @@
   <ItemGroup>
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="WebAPI\Controllers\ExportController.cs" />
     <Compile Include="WebAPI\Controllers\ModulesController.cs" />
     <Compile Include="WebAPI\Startup.cs" />
   </ItemGroup>
