+++ KInspector.Web/WebAPI/Controllers/ModulesController.cs
@@ -8,7 +8,9 @@ using System.Net.Http.Headers;
 using System.Web.Http;
 
 using Kentico.KInspector.Core;
+
 using Kentico.KInspector.Modules;
+using Kentico.KInspector.Modules.Export;
 
 namespace Kentico.KInspector.Web
 {
@@ -127,54 +129,5 @@ namespace Kentico.KInspector.Web
 				return Request.CreateResponse(HttpStatusCode.BadRequest, e.Message);
 			}
 		}
-
-        // GET api/modules/GetModulesResults
-        [ActionName("GetModulesResults")]
-        public HttpResponseMessage GetModulesResults([FromUri]IEnumerable<string> moduleNames, [FromUri]InstanceConfig config, [FromUri]string exportType = "xlsx")
-        {
-            ExportHelper.ExportType type;
-            if (!Enum.TryParse(exportType, true, out type))
-            {
-                throw new ArgumentNullException(nameof(exportType));
-            }
-
-            var instanceInfo = new InstanceInfo(config);
-            if (instanceInfo == null)
-            {
-                // TODO: Verify this
-                throw new ArgumentNullException(nameof(config));
-            }
-
-            if (moduleNames == null)
-            {
-                throw new ArgumentNullException(nameof(moduleNames));
-            }
-
-            try
-            {
-                var memoryStream = ExportHelper.GetExportStream(moduleNames, instanceInfo, type) as MemoryStream;
-                if (memoryStream == null)
-                {
-                    throw new Exception("Empty export file");
-                }
-
-                // Send stream
-                var response = Request.CreateResponse(HttpStatusCode.OK);
-                response.Content = new ByteArrayContent(memoryStream.ToArray());
-                response.Content.Headers.ContentType = new MediaTypeHeaderValue(ExportHelper.GetMimeType(type));
-                response.Content.Headers.ContentLength = memoryStream.Length;
-                response.Content.Headers.ContentDisposition = new ContentDispositionHeaderValue("attachment");
-                response.Content.Headers.ContentDisposition.FileName = $"KInspectorExport_{DateTime.Now.ToString("yyyyMMdd_HHmmss")}.{ExportHelper.GetExtension(type)}";
-                response.Headers.ConnectionClose = true;
-
-                memoryStream.Flush();
-
-                return response;
-            }
-            catch (Exception e)
-            {
-                return Request.CreateResponse(HttpStatusCode.InternalServerError, $"Error in processing modules. Error message: {e.Message}");
-            }
-        }
-	}
+    }
 }
